+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.1.3 (2025-06-25)
+* Chat history now supports images, downloads, and uploads.
+* Images supported by the code interpreter are no longer uploaded twice.
+* Convert `Chat.load` to a class method.
+* Code interpreter now replaces expired containers automatically.
+
 ## 0.1.2 (2025-06-19)
 * Add support for OpenAI's web search capability.
 * Add support for OpenAI's remote MCP capability.